@@ -1,3 +1,5 @@
 module github.com/cgrates/ltcache
 
-go 1.22
+go 1.24.1
+
+require golang.org/x/exp v0.0.0-20241004190924-225e2abe05e6
